+++ lib/models.dart
@@ -1,4 +1,8 @@
-export 'src/models/game.dart';
+export 'src/data/game.dart';
+export 'src/data/group.dart';
+export 'src/data/quiz.dart';
+export 'src/data/user.dart';
+export 'src/models/auth.dart';
 export 'src/models/group.dart';
 export 'src/models/quiz.dart';
 export 'src/models/user.dart';
