+++ lib/src/models/session.dart
@@ -1,0 +1,2 @@
+// Current quiz
+// List of users
