+++ lib/src/models/user.dart
@@ -1,56 +1,50 @@
-class User {
-  int _id;
-  int get id => _id;
-  String email;
-  String name;
+import 'dart:typed_data';
 
-  User(this._id, this.email, this.name);
+import 'package:flutter/cupertino.dart';
+import 'package:smart_broccoli/models.dart';
+import 'package:smart_broccoli/src/store/local/key_value.dart';
+import 'package:smart_broccoli/src/store/remote/user_api.dart';
 
-  factory User.fromJson(Map<String, dynamic> json) =>
-      User(json['id'], json['email'], json['name']);
+import 'auth.dart';
 
-  Map<String, dynamic> toJson() =>
-      <String, dynamic>{'id': id, 'email': email, 'name': name};
-}
+/// Class for making user profile requests
+/// For all methods in this class:
+/// `UnauthorisedRequestException` is thrown if the user is not logged in.
+/// `ForbiddenRequestException` is thrown if the user is logged in but not
+/// authorised to make the request.
+class UserStateModel with ChangeNotifier {
+  /// AuthModel object used to obtain token for requests
+  final AuthStateModel _authStateModel;
 
-/// User with login credientials (lecturer, coordinator)
-class RegisteredUser extends User {
-  /// To update the user's password, set this field then pass this object to
-  /// `UserModel.updateUser` to synchronise with server. This field is `null`.
-  String password;
+  UserApi _userApi;
 
-  RegisteredUser(int id, String email, String name) : super(id, email, name);
+  User _user;
+  User get user => _user;
 
-  factory RegisteredUser.fromJson(Map<String, dynamic> json) =>
-      RegisteredUser(json['id'], json['email'], json['name']);
+  Uint8List _profileImage;
 
-  Map<String, dynamic> toJson() {
-    Map map = super.toJson();
-    if (password != null) map['password'] = password;
-    return map;
+  KeyValueStore _keyValueStore;
+
+  /// Constructor for external use
+  UserStateModel(this._keyValueStore, this._authStateModel, {UserApi userApi}) {
+    _userApi = userApi ?? UserApi();
+    _user = User.fromJson(_keyValueStore.getItem('user'));
+    // TODO: load image
+  }
+
+  Future<void> refreshUser() async {
+    User user = await _userApi.getUser(_authStateModel.token);
+    _user = user;
+    _keyValueStore.setItem('user', user);
+    notifyListeners();
+  }
+
+  // TODO
+  Future<void> updateUser({String email, String password, String name}) {}
+
+  Future<void> getImage() async {
+    Uint8List image = await _userApi.getProfilePic(_authStateModel.token);
+    _profileImage = image;
+    // TODO: save to disk
   }
 }
-
-/// User without login credentials (student)
-class ParticipantUser extends User {
-  ParticipantUser(int id, {String email, String name}) : super(id, email, name);
-
-  factory ParticipantUser.fromJson(Map<String, dynamic> json) =>
-      ParticipantUser(json['id'], email: json['email'], name: json['name']);
-}
-
-class RegistrationException implements Exception {}
-
-/// Exception thrown when attempting to register (or update the profile of a
-/// user) with an already registered email
-class RegistrationConflictException extends RegistrationException {}
-
-/// Exception thrown when login is unsuccessful
-class LoginFailedException implements Exception {}
-
-/// Exception thrown when a participant user fails to register with the server
-class ParticipantJoinException extends RegistrationException {}
-
-/// Exception thrown when a user cannot be promoted to a registered user
-/// (e.g. the user is already registered)
-class ParticipantPromotionException extends RegistrationException {}
