+++ lib/src/models/user_repository.dart
@@ -1,0 +1,4 @@
+// Hold list of users
+// Handle profile pictures
+// Get user by id
+// Get new picture if picture ID has changed
